p5js/reaction-diffusion - Fix the datatype of reaction params
* This doesn't have any bearing until the params are passed in through the URL query string.
  - prior to this, the "integer" types would be passed into parseInt(...)

+++ p5js/coding-challenges/reaction-diffusion/classes/reaction_diffusion.js
@@ -57,12 +57,11 @@ class ReactionDiffusion {
       ];
   }
 
-
   optionsMetadata(){
     return [
-      { name: "diffRateA", type: "integer", default: 1}, 
-      { name: "diffRateB", type: "integer", default: 0.5}, 
-      { name: "killRate", type: "integer", default: 0.062}, 
+      { name: "diffRateA", type: "float", default: 1}, 
+      { name: "diffRateB", type: "float", default: 0.5}, 
+      { name: "killRate", type: "float", default: 0.062}, 
       { name: "feedRate", type: "float", default: 0.055}, 
       { name: "dt", type: "float", default: 0.05},
     ];
